pyproject.toml
[project]
name = "tool-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache[redis]>=0.12.3",
    "alembic>=1.17.1",
    "bcrypt>=5.0.0",
    "ddgs>=9.6.1",
    "fastapi>=0.120.0",
    "google-genai>=1.48.0",
    "httpx>=0.28.1",
    "instructor>=1.11.3",
    "jsonref>=1.1.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.12",
    "langfuse>=3.9.0",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.0",
    "omegaconf>=2.3.0",
    "passlib>=1.7.4",
    "plotly>=6.3.1",
    "psycopg-binary>=3.2.12",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "rich>=14.2.0",
    "slowapi>=0.1.9",
    "streamlit>=1.41.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "nb-black-formatter>=1.0.1",
    "prek>=0.2.12",
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]

[tool.ruff]
exclude = ["alembic", "alembic/**"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
